Rename Juju2Backend and move to module.
